1.37_dev3 Переключение режима очереди теперь работает Исправлена команда /stop Подсказки в /play теперь эффективней распределяют место в 100 символов Команды /playing, /jump теперь работают.

+++ commands/audio/remove.js
@@ -24,6 +24,7 @@ module.exports.slashBuilder = new SlashCommandBuilder()
   )
 
 module.exports.run = async ({ interaction, guild }) => {
+  if (!await AudioPlayer.playerIsExists(interaction)) return
   if (!await checkMemberInVoiceWithBotAndReply(interaction.member, interaction)) return
   await AudioPlayer.actions.deleteSongFromQueue(guild, interaction.options.get('position').value - 1, interaction.user.username)
 }
